Refactor the LPT Module
