+++ lib/lpt.rb
@@ -5,6 +5,7 @@ require "delegate"
 require "faraday"
 require "logger"
 
+require_relative "lpt/authentication"
 require_relative "lpt/version"
 require_relative "lpt/environment"
 require_relative "lpt/lpt_client"
@@ -32,10 +33,7 @@ module Lpt
   PREFIX_PROFILE = "LID"
 
   class << self
-    attr_accessor :api_username, :api_password, :merchant, :merchant_account,
-                  :entity, :proxy, :ca_store
-
-    attr_reader :api_version
+    include Lpt::Authentication
 
     attr_writer :open_timeout, :read_timeout, :write_timeout,
                 :max_network_retries, :initial_network_retry_delay,
@@ -54,6 +52,15 @@ module Lpt
 
     def api_version
       "v2"
+    end
+
+    def base_addresses(environment: nil)
+      if environment == Lpt::Environment::STAGING
+        { api_base: "staging-api-s2", cx_base: "cx.stg",
+          cx_api_base: "api.cx.stg" }
+      else
+        { api_base: "api", cx_base: "cx", cx_api_base: "api.cx" }
+      end
     end
 
     def open_timeout
@@ -92,44 +99,11 @@ module Lpt
       @log_level || LEVEL_DEBUG
     end
 
-    def client
-      assert_client_is_configured!
-      env = Lpt::Environment.factory(environment: environment)
-      Lpt::LptClient.factory(environment: env)
-    end
-
     protected
-
-    def assert_client_is_configured!
-      assert_username!
-      assert_password!
-      assert_merchant!
-      assert_entity!
-    end
 
     def assert_environment_is_valid!(env)
       msg = "Invalid Environment: #{env}"
       raise ArgumentError, msg unless envs.include?(env)
-    end
-
-    def assert_username!
-      msg = "Invalid API Username: #{api_username}"
-      raise ArgumentError, msg if api_username.blank?
-    end
-
-    def assert_password!
-      msg = "Invalid API Password"
-      raise ArgumentError, msg if api_password.blank?
-    end
-
-    def assert_merchant!
-      msg = "Invalid Merchant: #{merchant}"
-      raise ArgumentError, msg unless merchant.start_with? PREFIX_MERCHANT
-    end
-
-    def assert_entity!
-      msg = "Invalid Entity: #{entity}"
-      raise ArgumentError, msg unless entity.start_with? PREFIX_ENTITY
     end
 
     def standardize_environment(env)
